requirements.txt
# requirements.txt

flask
numpy
pandas
scikit-learn
tensorflow
keras
opencv-python
matplotlib
seaborn
joblib
google-generativeai
flask-cors
python-dotenv==1.0.0
 
# Security dependencies
bcrypt==4.0.1
email-validator==2.0.0
flask-limiter==3.5.0
werkzeug==2.3.7
requests==2.31.0
